chore: move enums
