+++ Editor/UI/Editors/AvatarImageConfigEditor.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Linq;
 using System.Reflection;
+using PlayerZero.Data;
 using UnityEditor;
 using UnityEditor.UIElements;
 using UnityEngine;
